Tweaked back prop. algo. added header and implem file for potential roll over optimizer idea. To be implemented.

+++ vector/vector.cpp
@@ -62,7 +62,7 @@ void Vector::printWeight() {
 
 void Vector::printHot() {
     std::cout << "Current Vector[hot]: ";
-    for (const int currentInteger : this->hotVector) {
+    for (const double currentInteger : this->hotVector) {
         std::cout << currentInteger << " ";
     }
     std::cout << std::endl;
